requirements.txt
# Django
Django==3.2.8

# Database
psycopg2===2.9.1

# Envs
django-dotenv==1.4.2
s3-environ==0.0.5

# Rest Framework
djangorestframework==3.12.2
django-filter==21.1

# Auth
django-rest-auth[with_social]==0.9.5

# Routers
drf-nested-routers==0.92.5

# Run Server
gunicorn==20.1.0
whitenoise==5.3.0

# Database
dj_database_url==0.5.0

# CORS
django-cors-headers==3.10.0

# Storage
django-storages==1.12.2

# Celery
celery==5.1.2
redis==3.5.3

# Sentry
sentry-sdk==1.4.3

# JSONSchema
jsonschema==4.1.2

# Django JSON Widget
django-json-widget==1.1.1
